% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authors.R
\name{authors}
\alias{authors}
\title{Get author information from a data object}
\usage{
authors(data_object)
}
\arguments{
\item{data_object}{An object produced by the function get_data().}
}
\value{
A data frame with project IDs, names of the authors of each data set, and their email.
}
\description{
Obtain author name and email from a data object downloaded from popler.
}
\examples{

# get author names and email from studies containing data from the genus Poa
poa_data = get_data(subset = genus == "Poa")
authors( poa_data )
}
