% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/close_DB_connection.R
\name{close_DB_connection}
\alias{close_DB_connection}
\title{Closes a connection with popler}
\usage{
close_DB_connection(connection)
}
\arguments{
\item{there}{is actually no parameter in this function!}
}
\description{
Close the connection with the popler database
}

