% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_unpack.R
\name{cov_unpack}
\alias{cov_unpack}
\title{Unpack the covariates for a data set}
\usage{
cov_unpack(x)
}
\arguments{
\item{An}{object produced by the function get_data(). NOTE: temporarily, data can come from 1 study only.}
}
\value{
A data frame whose columns represent the covariates of the data set.
}
\description{
Obtain author name and email from a data object downloaded from popler.
}
\examples{

# get author names and email from studies containing data from the genus Poa
demo_d <- get_data(proj_metadata_key == 8)
as.tbl( cov_unpack( demo_d ) )
}

