% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorsGroup.R
\name{errorsGroup}
\alias{errorsGroup}
\title{Informational message if no information is entered}
\usage{
errorsGroup(group_factors, possibleargs)
}
\arguments{
\item{group_factors}{factors}

\item{possible}{arguments that popler() can work with (internally defined)}
}
\value{
error messages, potentially
}
\description{
Service function for the popler() function
}

