% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorsTally.R
\name{errorsTally}
\alias{errorsTally}
\title{Error message regarding the tally_by arguments}
\usage{
errorsTally(tally_by, possibleargs)
}
\arguments{
\item{a}{vector of column names}

\item{possible}{arguments that popler() can work with (internally defined)}
}
\value{
an error message, if one of the vector elements is misspelled
}
\description{
Service function for the popler() function
}

