% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fancy_browse.R
\name{fancy_browse}
\alias{fancy_browse}
\title{A user-friendly description of the browse() metadata}
\usage{
fancy_browse(input, md_file = "./browse.Rmd", html_file = "./browse.html")
}
\arguments{
\item{input}{a popler object returned from browse() or get_data()}

\item{md_file}{Specify the filename and location for the generated markdown file (optional)}

\item{html_file}{Specify the filename and location for the generated html file (optional)}
}
\description{
Provides information on the columns of metadata contained in the popler database, and the kind of data contained in those columns. Also provides citations and example code for the object.
}
\examples{
# Full dictionary
search <- browse(community=="no" & duration_years > 15)
fancy_browse(search)

data <- get_data(search)
fancy_browse(data) # same as above

}

