% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fancy_dictionary.R
\name{fancy_dictionary}
\alias{fancy_dictionary}
\title{A user-friendly dictionary of the popler metadata}
\usage{
fancy_dictionary(full_tbl = FALSE, md_file = "./dictionary.Rmd",
  html_file = "./dictionary.html")
}
\arguments{
\item{full_tbl}{Should the function return the standard columns, or the full main table?}

\item{md_file}{Specify the filename and location for the generated markdown file (optional)}

\item{html_file}{Specify the filename and location for the generated html file (optional)}
}
\description{
Provides information on the columns of metadata contained in the popler database, and the kind of data contained in those columns.
}
\examples{
# Full dictionary
fancy_dictionary(full_tbl = TRUE)

# "Abridged" version
fancy_dictionary()
}

