% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_methods.R
\name{filter.browse}
\alias{filter.browse}
\alias{filter.get_data}
\title{Methods for dplyr verbs}
\usage{
\method{filter}{browse}(.data, ...)

\method{filter}{get_data}(.data, ...)
}
\arguments{
\item{.data}{A \code{browse} or \code{get_data} object}

\item{...}{logical conditions}
}
\description{
Subsets a \code{browse} or \code{get_data} object based on
logical statements.
}
