% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatMainTable.R
\name{formatMainTable}
\alias{formatMainTable}
\title{Formats the main table of popler()}
\usage{
formatMainTable(x)
}
\arguments{
\item{main}{table of popler()}
}
\value{
formatted popler() main table
}
\description{
This function converts every column encoded as "factor" into "character"
}

