% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Browse the popler database}
\usage{
get_data(browsed_data = NULL, subset = NULL, add_columns = NULL,
  subtract_columns = NULL, metadata = FALSE)
}
\arguments{
\item{browsed_data}{A on object produced by the function browse()}

\item{subset}{A logical argument A string that to specify which columns to add in the default columns used in queries to the database. See Details for default columns.}

\item{add_columns}{A string to specify which columns the user wants to add in queries to the database}

\item{subtract_columns}{A string to specify which, among the default columns, the user wishes to discard in queries to the database}

\item{metadata}{Should the function open the webpage containing the metadata for each one of the projects downloaded?}
}
\value{
A data frame of the selected data.
}
\description{
This function browses LTER studies contained in the popler database. The user can group, tally, and subset the data sets based on a number of the factors (or columns) of the database.
}
\examples{

# browse a study, then get the data associated with it
grasshop = browse(proj_metadata_key == 21)
gh_data = get_data(grasshop)
# further subset this data set, based on year
gh_data_96_99 = get_data(grasshop, year > 1995 & year < 2000)
}

