% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyword.R
\name{\%=\%}
\alias{\%=\%}
\title{keyword binary operator}
\usage{
var \%=\% pattern
}
\arguments{
\item{var}{The name of a vector object.}

\item{keyword}{A string containing the keyword(s) that need be matched}
}
\value{
A logical vector.
}
\description{
Match one or more string patterns with the contents of a string vector.
This is a wrapper of the grepl() function in base.
}
\examples{
letters \%=\% "a"

# matching is case insensitive 
letters \%=\% "A"

# Matches multiple patterns
letters \%=\% c("A","c","d")
}

