% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infoMessage.R
\name{infoMessage}
\alias{infoMessage}
\title{Error message regarding the group_factors arguments}
\usage{
infoMessage(group_factors, tally_by, criteria, possibleargs)
}
\arguments{
\item{group_factors}{factors}

\item{possibleargs}{possible arguments}
}
\value{
error messages, potentially
}
\description{
Service function for the popler() function
}

