% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_webpage.R
\name{metadata_webpage}
\alias{metadata_webpage}
\title{Open web browser with the metadata webpage}
\usage{
metadata_webpage(study_id = NULL)
}
\arguments{
\item{study_id}{The id of the study. Works with one id at a time.}
}
\description{
Opens the metadata page of study, using its id
}
\examples{
look up metadata link for study number 21
metadata_webpage(study_id=21)
}

