% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleColumns.R
\name{multipleColumns}
\alias{multipleColumns}
\title{Convert entries of popler function}
\usage{
multipleColumns(x)
}
\arguments{
\item{a}{vector of column names}
}
\value{
a vector of column names
}
\description{
Service function for the popler() function
}

