% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_methods.R
\name{mutate.browse}
\alias{mutate.browse}
\alias{mutate.get_data}
\title{Methods for dplyr verbs}
\usage{
\method{mutate}{browse}(.data, ...)

\method{mutate}{get_data}(.data, ...)
}
\arguments{
\item{.data}{A \code{browse} or \code{get_data} object}

\item{...}{Name-value pairs of expressions. Use \code{NULL} to drop a 
variable.}
}
\description{
Add new columns to a \code{browse} or \code{get_data} object.
}
