% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_DB_connection.R
\name{open_DB_connection}
\alias{open_DB_connection}
\title{Opens a connection with popler}
\usage{
open_DB_connection()
}
\arguments{
\item{there}{is actually no parameter in this function!}
}
\description{
Opens a connection with the popler database
}

