% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_to_sql_search.R
\name{parse_to_sql_search}
\alias{parse_to_sql_search}
\title{Translate R logical statements into SQL}
\usage{
parse_to_sql_search(..., lazy = T)
}
\arguments{
\item{...}{A logical expression that specifies which records to keep.}

\item{lazy}{logical argument: lazy evaluation of logical expression or not?}
}
\value{
A string with an SQL statement
}
\description{
USe dplyr translate_sql() to translate R logical statements into SQL
}

