% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popler.R
\name{popler}
\alias{popler}
\title{Browse the popler database}
\usage{
popler(group_factors = NULL, tally_by = NULL, criteria = NULL,
  trim = TRUE)
}
\arguments{
\item{tally_by}{A vector of characters that specifies which fields to tally}

\item{criteria}{A logical expression that specifies which records to keep.
Synthax is the same as that used in the base function subset.}

\item{trim}{should strings be truncated at the 50th character? Default is TRUE.}

\item{group_by}{A vector of characters that specifies the factors by which data should be grouped.}
}
\value{
A data frame reporting groping factors and tallies, based on the subsetting criteria.
}
\description{
This function browses LTER studies contained in the popler database. The user can group, tally, and subset the data sets based on a number of the factors (or columns) of the database.
}
\examples{
#returns all of the species included in the data set
popler(group_by="species")

#returns the number of species species for each study
popler(group_by="title",tally_by="species")

#returns the studies in which that contain the genus "Abietinaria"
popler(group_by="title",criteria=genus=="Abietinaria")
}

