% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{pplr_browse}
\alias{pplr_browse}
\title{Browse the metadata of projects contained in the popler database}
\usage{
pplr_browse(..., full_tbl = FALSE, vars = NULL, trim = TRUE,
  view = FALSE, keyword = NULL, report = FALSE)
}
\arguments{
\item{...}{A logical expression to subset popler's main table}

\item{full_tbl}{Should the function return the standard 
columns, or the full main table?}

\item{vars}{A vector of characters: which variables 
of popler's main table should be selected?}

\item{trim}{If TRUE, strings are truncated at the 50th character.
Default is TRUE.}

\item{view}{If TRUE, opens up a spreadsheet-style data viewer.}

\item{keyword}{A string that selects}

\item{report}{If TRUE, function produces a markdown 
report about each study's metadata, and opens it as a html page.}
}
\value{
A data frame combining the metadata of each project 
and the taxonomic units associated with each project.

This data frame is of class \code{popler}, \code{data.frame},
\code{tbl_df}, and \code{tbl}.
}
\description{
pplr_browse() reports the metadata of LTER studies contained in the popler database. 
The user can subset what data, and which variables to visualize.
}
\examples{

\dontrun{
# No arguments return the standard 16 columns of popler's main table
default_vars = pplr_browse()

# full_tbl = TRUE returns the full table
all_vars = pplr_browse(full_tbl = TRUE)

# subset only data from the sevilleta LTER, and open the relative report in a html page
sev_data = pplr_browse(lterid == "SEV", report = TRUE)

# consider only plant data sets 
plant_data = pplr_browse(kingdom == "Plantae")

# Select only the data you need
three_columns = pplr_browse(vars = c("title","proj_metadata_key","genus","species"))

# Select only the data you need
study_21 = pplr_browse( proj_metadata_key == 25)

# Select studies that contain word "parasite"
parasite_studies = pplr_browse( keyword = "parasite")
}

}
