% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popler_citation.R
\name{pplr_citation}
\alias{pplr_citation}
\title{Provide citations for a popler browse or get_data object.}
\usage{
pplr_citation(input, bibtex_path = NULL)
}
\arguments{
\item{input}{a popler object returned from browse() or get_data()}

\item{bibtex_path}{Specify the filename and location for 
the generated markdown file (optional)}
}
\value{
A list of references from \code{input}
}
\description{
Returns a bibliography, Bibtex citations, and an acknowledgement template.
}
\examples{
\dontrun{
# make a browse object
metadata <- pplr_browse(proj_metadata_key \%in\% c(17, 317, 494))

# cite the projects
cite <- pplr_citation(metadata)

# cite$bibliography          # the bibliography
# cite$Bibtex                # Bibtex entries for each dataset
# cite$acknowledgement       # acknowledgement template
}

}
