% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{pplr_dictionary}
\alias{pplr_dictionary}
\title{Dictionary of the popler metadata variables}
\usage{
pplr_dictionary(..., full_tbl = FALSE)
}
\arguments{
\item{...}{A sequence of (unquoted) variables specifying one
or more variables of popler's main table for which dictionary 
information is needed}

\item{full_tbl}{logical; If \code{TRUE}, the function
returns a table describing the variables of the full main table.
If \code{FALSE}, the function returns a table describing the standard 
variables. Default is \code{FALSE}.}
}
\description{
Describes the metadata variables contained
in the popler database, and shows their content.
}
\examples{
\dontrun{
# Column names
column_names <- pplr_dictionary(full_tbl = FALSE)

# Dictionary information
dictionary_lter <- pplr_dictionary(lterid, full_tbl = FALSE)

# multiple columns
dictionary_lter_lat <- pplr_dictionary(lterid,lat_lter, full_tbl = FALSE)
}
}
