% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_url.R
\name{pplr_metadata_url}
\alias{pplr_metadata_url}
\title{Get metadata information from a data object}
\usage{
pplr_metadata_url(input)
}
\arguments{
\item{input}{An object produced by the function \code{pplr_get_data()}.}
}
\description{
Load the webpage containing the metadata of the data sets 
contained in objects produced by \code{pplr_browse} or downloaded through 
\code{pplr_get_data()}. If you downloaded data from multiple projects, 
this function will open multiple webpages. This is a wrapper of function 
\code{browseURL} in \code{base}.
}
\examples{

\dontrun{
# Load the metadata webpages of the projects that contain data from the Poa genus.
fes_d <- pplr_browse(genus == "Festuca")
pplr_metadata_url( fes_d )
}

}
