% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_dictionary.R
\name{pplr_report_dictionary}
\alias{pplr_report_dictionary}
\title{A user-friendly dictionary of the popler metadata}
\usage{
pplr_report_dictionary(full_tbl = FALSE, md_file = NULL, html_file = NULL)
}
\arguments{
\item{full_tbl}{Should the function return the standard 
columns, or the full main table?}

\item{md_file}{Specify the filename and location for 
the generated markdown file (optional)}

\item{html_file}{Specify the filename and location for the 
generated html file (optional)}
}
\value{
This function is called for its side effects and does not 
return anything.
}
\description{
Provides information on the columns of metadata contained 
in the popler database, and the kind of data contained in those columns.
}
\examples{
\dontrun{
# Full dictionary
pplr_report_dictionary(full_tbl = TRUE)

# "Abridged" version
pplr_report_dictionary()
}

}
