% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_metadata.R
\name{pplr_report_metadata}
\alias{pplr_report_metadata}
\title{Open a report of the metadata of project(s) as an html page}
\usage{
pplr_report_metadata(input, md_file = "./browse.Rmd",
  html_file = "./browse.html")
}
\arguments{
\item{input}{A popler object returned from \code{pplr_browse()} or 
\code{pplr_get_data()}}

\item{md_file}{Specify the filename and location for the generated markdown
file (optional)}

\item{html_file}{Specify the filename and location for
the generated html file (optional)}
}
\value{
An invisible copy of \code{input}
}
\description{
Generates a readable report of the metadata describing data sets contained in
popler. The report contains citations, the links to the original URL of each 
data set, and example code to obtain the metadata and data of the projects
represented in the html page.
}
\examples{
\dontrun{
# Full dictionary
one_spp <- pplr_browse(community == "no" & duration_years > 15)
pplr_report_metadata(one_spp)

data <- pplr_get_data(one_spp)
pplr_report_metadata(data) # same as above
}
}
