% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_replication.R
\name{pplr_site_rep}
\alias{pplr_site_rep}
\title{Spatial-temporal replication of data sets}
\usage{
pplr_site_rep(input, return_plot = FALSE)
}
\arguments{
\item{input}{The output from \code{pplr_get_data()}. This will not work
with output from \code{pplr_browse} as the raw data is required to calculate
the amount of replication.}

\item{return_plot}{A logical indicating whether to return a copy of the 
\code{input} data or the \code{ggplot} object created by the function. Use
\code{TRUE} to return the \code{ggplot} object for subsequent modification.
Use \code{FALSE} to return an invisible copy of the \code{input} object (
useful for piping). Default is \code{FALSE}.}
}
\value{
The \code{input} object (invisibly) or a \code{ggplot2} object.
}
\description{
Create figures for temporal replication of sites in a given 
data set.
}
\examples{
\dontrun{

library(ggplot2)
library(dplyr)

# create an unmodified figure
BNZ <- pplr_get_data(lterid == 'BNZ')

pplr_site_rep(BNZ)

# Return the figure instead of the data for subsequent modification
Antarctica <- pplr_get_data(lterid == 'PAL')

pplr_site_rep(Antarctica,
              return_plot = TRUE) + 
   ggtitle("Penguins Rock!")
   
# Use within pipes. Cannot return and modify the figure this way.
pplr_get_data(lterid == 'SEV') \%>\% 
  pplr_site_rep(return_plot = FALSE) \%>\%
  pplr_report_metadata()
}

}
