% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{pplr_summary}
\alias{pplr_summary}
\title{search}
\usage{
pplr_summary(limit = 10, offset = 0, ...)
}
\arguments{
\item{limit}{number of records to return, default: 10}

\item{offset}{record number to start at, default: first record}

\item{...}{curl options passed on to [crul::HttpClient]}
}
\description{
search
}
\examples{
# basic example
pplr_summary()
# pass in curl options for debugging, seeing http request details
pplr_summary(verbose = TRUE)
}
