% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{pplr_summary_table_check}
\alias{pplr_summary_table_check}
\title{Check last time summary_table was updated}
\usage{
pplr_summary_table_check()
}
\description{
Checks the main table's age. If it's more than 6 weeks old, 
returns a message suggesting an update.
}
\seealso{
\code{\link{pplr_summary_table_update}}
}
