% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{pplr_summary_table_update}
\alias{pplr_summary_table_update}
\title{Update \code{popler}'s summary table}
\usage{
pplr_summary_table_update()
}
\value{
This function is called for its side effect and does not return 
anything
}
\description{
Automatically retrieve most up to date version of \code{popler}
summary table
}
\note{
The \code{summary_table} is often called internally by popler functions,
 but can also be accessed directly by calling \code{pplr_summary_table_import()}.
}
\seealso{
\code{\link{pplr_summary_table_check}}
}
