% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{rebrowse.browse}
\alias{rebrowse.browse}
\title{Rebrowse a \code{browse} object}
\usage{
\method{rebrowse}{browse}(popler, ...)
}
\arguments{
\item{popler}{An object of class \code{popler} and either \code{browse} or
\code{get_data}.}

\item{...}{unused}
}
\description{
Rebrowse a \code{browse} object
}
