% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{rebrowse.get_data}
\alias{rebrowse.get_data}
\title{Rebrowse a \code{get_data} object}
\usage{
\method{rebrowse}{get_data}(popler, ...)
}
\arguments{
\item{popler}{An object of class \code{popler} and either \code{browse} or
\code{get_data}.}

\item{...}{unused}
}
\description{
Rebrowse a \code{get_data} object
}
