% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{rebrowse}
\alias{rebrowse}
\title{Take a \code{browse} object and update it with \code{full_tbl = TRUE} & 
\code{trim = FALSE}.}
\usage{
rebrowse(popler, ...)
}
\arguments{
\item{popler}{An object of class \code{popler} and either \code{browse} or
\code{get_data}.}

\item{...}{unused}
}
\description{
Take a \code{browse} object and update it with \code{full_tbl = TRUE} & 
\code{trim = FALSE}.
}
