% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectByCriteria.R
\name{selectByCriteria}
\alias{selectByCriteria}
\title{Select which records to keep from the popler main table.}
\usage{
selectByCriteria(x, criteria)
}
\arguments{
\item{The}{popler main table}

\item{A}{logical expression that specifies which records to keep.}
}
\value{
an error message, if one of the vector elements is misspelled
}
\description{
Service function for the popler() function
}

