% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_SSH.R
\name{start_SSH}
\alias{start_SSH}
\title{Starts tunnelling with popler}
\usage{
start_SSH()
}
\arguments{
\item{there}{is actually no parameter in this function!}
}
\description{
Start tunneling with the popler database itself
}

