% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_SSH.R
\name{stop_SSH}
\alias{stop_SSH}
\title{Stops tunnelling with popler}
\usage{
stop_SSH()
}
\arguments{
\item{there}{is actually no parameter in this function!}
}
\description{
Stop tunneling with the popler database
}

