% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_names.R
\name{study_names}
\alias{study_names}
\title{Print popler study names in a legible form.}
\usage{
study_names(browsed, width = 60)
}
\arguments{
\item{browsed}{An object created by the function browse()}

\item{width}{A positive integer giving the target column for wrapping lines in the output (from base function strwrap)}
}
\value{
prints the names of popler studies as separate paragraphs
}
\description{
Prints the names of popler studies as paragraphs of known width. This is a wrapper of the strwrap function in base.
}
\examples{

# Look up names
study_names(browse())
study_names(browse(datatype == "individual"))
study_names(browse(datatype == "individual"), width = 30)
}

