% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{summary_table_check}
\alias{summary_table_check}
\title{Checks if the summary table needs to be updated}
\usage{
summary_table_check()
}
\description{
Checks the main table's age. If it's more than 6 weeks old, 
returns a message suggesting an update
}
\seealso{
\code{\link{summary_table_update}}
}
