% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{summary_table_import}
\alias{summary_table_import}
\title{Imports the \code{summary_table} object to R}
\usage{
summary_table_import()
}
\value{
A table summarizing each type of data in the popler data base.
}
\description{
Imports the \code{summary_table} object to the top level environment
when called by the user.
}
\details{
This function is often called internally to load the summary table
and extract some metadata that is used in that functional context. Loading
summary table may be useful for exploring what data is in popler in case
the vignettes prove insufficiently graphic. However, be aware
it is loaded by functions each time they are called, so any manipulations
by the user at the top level will not be passed to those functions.
}
