% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tall.R
\name{tall}
\alias{tall}
\title{Tally records from the popler main table}
\usage{
tall(subsetDat, tally_by, columnNames, trim)
}
\arguments{
\item{The}{popler main table}

\item{A}{logical expression that specifies which records to keep.}
}
\value{
an error message, if one of the vector elements is misspelled
}
\description{
Service function for the popler() function
}

