% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tally_by.R
\name{tally_by}
\alias{tally_by}
\title{Tally the number of instances in the popler database}
\usage{
tally_by(browse_data, tally_columns = "species", group_factors = NULL,
  trim = TRUE)
}
\arguments{
\item{trim}{if TRUE, strings are truncated at the 50th character.}

\item{x}{A data frame produced by the browse() function}

\item{tally_column}{Should the function return the standard columns, or the full main table?}

\item{group_factor}{grouping factor by which to count instances.}
}
\value{
A data frame showing title's, or a subset thereof
}
\description{
tally_by() tallies instances (e.g. number of species), in the popler database
Instances can be counted by themselves, or using grouping factors (e.g. species by study)
}
\examples{
# Tallies without grouping factor
tally_by(browse(), tally_columns = "title", trim = TRUE)
         
# Number of species by study
tally_by(browse(), tally_columns = "species",
         group_factors = "title", trim = TRUE)
         
# Number of studies by LTER site
tally_by(browse(), tally_columns = "title",
         group_factors = "lterid", trim = TRUE)
}

