% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimDisplay.R
\name{trimDisplay}
\alias{trimDisplay}
\title{Trim the display of the popler() output}
\usage{
trimDisplay(x)
}
\arguments{
\item{a}{data frame}
}
\value{
a data frame whose string elements are trimmed at the 50th character
}
\description{
Service function for the popler() function
}

