% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tunnel_query.r
\name{tunnel_query}
\alias{tunnel_query}
\title{Browse the popler database}
\usage{
tunnel_query(...)
}
\arguments{
\item{A}{logical expression using R synthax.}
}
\value{
A data frame of the selected data.
}
\description{
This function browses LTER studies contained in the popler database. The user can group, tally, and subset the data sets based on a number of the factors (or columns) of the database.
}
\examples{
#returns ALL DATA referring to "Abietinaria"
get_fxn(genus=="Abietinaria")

#returns data from the study number 7
get_fxn(medarecordid==7)
}

