% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tunnel_query_.R
\name{tunnel_query_}
\alias{tunnel_query_}
\title{Browse the popler database, without lazy evaluation}
\usage{
tunnel_query_(...)
}
\arguments{
\item{A}{logical expression using R synthax.}
}
\value{
A data frame of the selected data.
}
\description{
This function fetches LTER studies contained in the popler database without lazy evaluation. The user can group, tally, and subset the data sets based on a number of the factors (or columns) of the database.
}
\examples{
#returns ALL DATA referring to "Abietinaria"
get_fxn(genus=="Abietinaria")

#returns data from the study number 7
get_fxn(medarecordid==7)
}

