% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniqueValues.R
\name{uniqueValues}
\alias{uniqueValues}
\title{Find unique values from the popler main table}
\usage{
uniqueValues(subsetDat, columnNames, trim)
}
\arguments{
\item{A}{selection of, or the entire popler main table.}

\item{A}{logical expression that specifies which records to keep.}
}
\value{
an error message, if one of the vector elements is misspelled
}
\description{
Service function for the popler() function
}

