% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{browse}
\alias{browse}
\title{Browse the project metadata contained in the popler database}
\usage{
browse(..., full_tbl = FALSE, vars = NULL, trim = TRUE, view = FALSE,
  keyword = NULL)
}
\arguments{
\item{...}{A logical expression to subset popler's main table}

\item{full_tbl}{Should the function return the standard columns, or the full main table?}

\item{vars}{A vector of characters: which variables of popler's main table should be selected?}

\item{trim}{If TRUE, strings are truncated at the 50th character. Default is TRUE.}

\item{view}{If TRUE, opens up a spreadsheet-style data viewer.}

\item{keyword}{A string that selects}
}
\value{
A data frame combining the metadata of each project and the taxonomic units associated with each project.

This data frame is of class "popler", "data.frame", "tbl_df", and "tbl".
}
\description{
browse() reports the metadata of LTER studies contained in the popler database. 
The user can subset what data and what columns to visualize.
}
\examples{
# No arguments return the standard 16 columns of popler's main table
default_vars = browse()

# full.table==T returns the full table
all_vars = browse(full_tbl = TRUE)

# subset only data from the sevilleta LTER 
sev_data = browse(lterid == "SEV")

# consider only plant data sets 
plant_data = browse(kingdom == "Plantae")

# Select only the data you need
three_columns = browse(vars = c("title","proj_metadata_key","genus","species"))

# Select only the data you need
study_21 = browse( proj_metadata_key == 21)

# More straighforward using keyword operator, \%=\%
parasite_npp = browse( title \%=\% "parasit", trim = FALSE)
}
