% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{browse}
\alias{browse}
\title{Browse the popler database}
\usage{
browse(select_columns = NULL, full.table = F)
}
\arguments{
\item{select_columns}{A vector of characters that specifies which columns of popler's main table should be selected.}

\item{full.table}{Should the function return the standard columns, or the full main table?}
}
\value{
A data frame of the main table, or a subset thereof
}
\description{
This function browses LTER studies contained in the popler database. The user can group, tally, and subset the data sets based on a number of factors (or columns) of that characterize datasets in the database.
}
\examples{
# No arguments return the standard 16 columns of popler's main table
browse()
# full.table==T returns the full table
browse(full.table = T)
# Select only the data you need
browse(c("title","genus","species"), full.table = T)
}

