% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{dictionary}
\alias{dictionary}
\title{Dictionary of the popler metadata}
\usage{
dictionary(..., full_tbl = FALSE)
}
\arguments{
\item{...}{A sequence of (unquoted) variables specifying one or more columns of popler's main table for which dictionary information is needed}

\item{full_tbl}{Should the function return the standard columns, or the full main table?}
}
\description{
Provides information on the columns of metadata contained in the popler database, and the kind of data contained in those columns.
}
\examples{
# Column names
column_names <- dictionary(full_tbl = FALSE)

# Dictionary information
dictionary_lter <- dictionary(lterid, full_tbl = FALSE)

# multiple columns
dictionary_lter_lat <- dictionary(lterid,lat_lter, full_tbl = FALSE)
}
