% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Browse the popler database}
\usage{
get_data(..., add_vars = NULL, subtract_vars = NULL)
}
\arguments{
\item{...}{A list of one or two objects: an object produced by browse, a logical statement, or both.}

\item{add_vars}{A string to specify which columns the user wants to add in queries to the database}

\item{subtract_vars}{A string to specify which, among the default columns, the user wishes to discard in queries to the database}
}
\value{
A data frame of the selected data.
}
\description{
This function browses LTER studies contained in the popler database. The user can group, tally, and subset the data sets based on a number of the factors (or columns) of the database.
}
\examples{

# browse a study, then get the data associated with it
grasshop = browse(proj_metadata_key == 21)
gh_data = get_data(grasshop)

# further subset this data set, based on year
gh_data_96_99 = get_data(grasshop, year > 1995 & year < 2000)

# insect data sets from the SEV lter site
insect_sev = browse(class == "Insecta" & lterid == "SEV")
insect_21_yrs96_99 = get_data(insect_sev, year > 1995 & year < 2000 & proj_metadata_key == 21)

insect_21_25 = get_data((proj_metadata_key == 21 | proj_metadata_key == 25) & year < 1995 )
}

