% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyword.R
\name{\%==\%}
\alias{\%==\%}
\title{keyword binary operator}
\usage{
var \%==\% pattern
}
\arguments{
\item{var}{A variable of contained in popler's metadata}

\item{keyword}{The keyword to be matched}
}
\value{
A logical vector.
}
\description{
Search for elements of metadata variables in popler which match a keywords.
}
\examples{
# No arguments return the standard 16 columns of popler's main table
letters \%==\% "a"

# Performs case insensitive matching
letters \%==\% "A"

# Mataches multiple patterns
letters \%==\% c("A","c","d")

}

