% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyword.R
\name{\%=\%}
\alias{\%=\%}
\title{keyword binary operator}
\usage{
var \%=\% keyword
}
\arguments{
\item{var}{The name of a character vector.}

\item{keyword}{A character string containing a keyword to be matched to the argument `var`}
}
\value{
A logical vector.
}
\description{
Match one string pattern with the contents of a string vector.
This is a wrapper of the grepl() function in base.
}
\examples{
letters \%=\% "a"

# matching is case insensitive 
letters \%=\% "A"
}

