% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_unpack.R
\name{pplr_cov_unpack}
\alias{pplr_cov_unpack}
\title{Unpack the covariates contained in the dataset downloaded via 
\code{pplr_get_data()}}
\usage{
pplr_cov_unpack(input)
}
\arguments{
\item{input}{An object of class \code{get_data}.}
}
\value{
A data frame whose columns represent the covariates of the dataset
downloaded via \code{pplr_get_data()}. Note that these covariates are 
contained in the \code{covariates} column datasets downloaded using
\code{pplr_get_data()}.
}
\description{
Create a data frame by "extracting" the \code{covariates} column
contained in an dataset downloaded with \code{pplr_get_data()}.
}
\examples{
\dontrun{
library(dplyr)
demo_d <- pplr_get_data(proj_metadata_key == 8)
as.tbl( pplr_cov_unpack( demo_d ) )
}

}
