% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_funs.R
\name{pplr_maps}
\alias{pplr_maps}
\title{Generate maps of LTER sites}
\usage{
pplr_maps(input, return_maps = FALSE)
}
\arguments{
\item{input}{An object created by either \code{pplr_browse()} or
\code{pplr_get_data()}}

\item{return_maps}{A logical indicating whether to return the \code{ggplot}
objects or an invisible copy of the input data. Use \code{FALSE} to retain
the ability to use this in a pipe. Use \code{TRUE} if you want to modify
the \code{ggplot} objects interactively.}
}
\value{
The \code{input} object (invisibly).
}
\description{
Generates maps of LTER sites in a given \code{input} object.
Sizes of site markers correspond to the number of studies at a given site.
}
