% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_url.R
\name{pplr_metadata_url}
\alias{pplr_metadata_url}
\title{Get metadata information from a data object}
\usage{
pplr_metadata_url(data_object)
}
\arguments{
\item{data_object}{An object produced by the function pply_get_data()}
}
\description{
Load the webpage containing the metadata of the data sets downloaded through 
get_data(). If you downloaded data from multiple projects, this function opens multiple webpages. 
This is a wrapper of function browseURL in base.
}
\examples{

\dontrun{
# Load the metadata webpages of the projects that contain data from the Poa genus.
fes_d <- pplr_browse(genus == "Festuca")
pplr_metadata_url( fes_d )
}

}
