% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_dictionary.R
\name{pplr_report_dictionary}
\alias{pplr_report_dictionary}
\title{A user-friendly dictionary of the popler metadata}
\usage{
pplr_report_dictionary(full_tbl = FALSE, md_file = NULL,
  html_file = NULL)
}
\arguments{
\item{full_tbl}{logical; if \code{TRUE} function returns the variables 
contained in the full main table. If \code{FALSE}, functions returns only the
standard variables. Default is \code{FALSE}.}

\item{md_file}{Specify the filename and location for 
the generated markdown file (optional)}

\item{html_file}{Specify the filename and location for the 
generated html file (optional)}
}
\value{
This function is called for its side effects and does not 
return an R object.
}
\description{
Provides information on the variables of metadata contained in the popler 
database, and the kind of data contained in those variables.
}
\examples{
\dontrun{
# Full dictionary
pplr_report_dictionary(full_tbl = TRUE)

# "Abridged" version
pplr_report_dictionary()
}

}
