% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{rebrowse}
\alias{rebrowse}
\alias{rebrowse.browse}
\alias{rebrowse.get_data}
\title{Take a \code{browse} object and update it with \code{full_tbl = TRUE}.}
\usage{
rebrowse(input, ...)

\method{rebrowse}{browse}(input, ...)

\method{rebrowse}{get_data}(input, ...)
}
\arguments{
\item{input}{An object of class either \code{browse} or
\code{get_data}.}

\item{...}{unused}
}
\description{
Take a \code{browse} object and update it with \code{full_tbl = TRUE}.
}
