% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{summary_table_update}
\alias{summary_table_update}
\title{Automaticcaly retrieve most up to date version of \code{popler}
summary table}
\usage{
summary_table_update()
}
\value{
Updates the data object \code{summary_table}.
}
\description{
Automaticcaly retrieve most up to date version of \code{popler}
summary table
}
\note{
This object is often called internally by popler functions,
 but can also be accessed by calling \code{data(summary_table)}.
}
\seealso{
\code{\link{summary_table_check}}
}
