% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tunnel_query.r
\name{tunnel_query}
\alias{tunnel_query}
\title{Browse the popler database}
\usage{
tunnel_query(..., add_columns = NULL, subtract_columns = NULL,
  metadata = T)
}
\arguments{
\item{...}{A logical expression using R synthax.}

\item{add_columns}{A string that to specify which columns to add in the default columns used in queries to the database. See Details for default columns.}

\item{subtract_columns}{A string to specify which columns, among the default columsn, the user wishes to discard in queries to the database}
}
\value{
A data frame of the selected data.
}
\description{
This function browses LTER studies contained in the popler database. The user can group, tally, and subset the data sets based on a number of the factors (or columns) of the database.
}
\examples{
#returns ALL DATA referring to "Abietinaria"
get_fxn(genus=="Abietinaria")

#returns data from the study number 7
get_fxn(medarecordid==7)
}

